Use vector slice destructuring to read values
Instead of using the awkward self.value.get(X) calls to get the values
for the record types that need them (MX and SRV), use a match on
self.values.as_slice() to make sure we have the exact number of elements
we need. This also leaves us with named variables that can be used when
printing errors, and just overall more readable code (although the
parsing is still a bit awkward).